+++ packages/loader/container-loader/src/container.ts
@@ -42,8 +42,6 @@ import {
 import {
     BlobCacheStorageService,
     readAndParse,
-    OnlineStatus,
-    isOnline,
     ensureFluidResolvedUrl,
     combineAppAndProtocolSummary,
     readAndParseFromBlobs,
@@ -83,7 +81,6 @@ import {
     ChildLogger,
     EventEmitterWithErrorHandling,
     PerformanceEvent,
-    raiseConnectedEvent,
     TelemetryLogger,
     connectedEventName,
     disconnectedEventName,
@@ -91,15 +88,15 @@ import {
 import { Audience } from "./audience";
 import { ContainerContext } from "./containerContext";
 import { debug } from "./debug";
-import { IConnectionArgs, DeltaManager, ReconnectMode } from "./deltaManager";
+import { IConnectionArgs, DeltaManager } from "./deltaManager";
 import { DeltaManagerProxy } from "./deltaManagerProxy";
 import { Loader, RelativeLoader } from "./loader";
 import { pkgVersion } from "./packageVersion";
 import { parseUrl, convertProtocolAndAppSummaryToSnapshotTree } from "./utils";
+import { connectEventName, ConnectionState, ConnectionStateHandler } from "./connectionStateHandler";
 
 const detachedContainerRefSeqNumber = 0;
 
-const connectEventName = "connect";
 const dirtyContainerEvent = "dirty";
 const savedContainerEvent = "saved";
 
@@ -142,23 +139,6 @@ export interface IContainerConfig {
      */
     clientDetailsOverride?: IClientDetails;
     id?: string;
-}
-
-export enum ConnectionState {
-    /**
-     * The document is no longer connected to the delta server
-     */
-    Disconnected,
-
-    /**
-     * The document has an inbound connection but is still pending for outbound deltas
-     */
-    Connecting,
-
-    /**
-     * The document is fully connected
-     */
-    Connected,
 }
 
 /**
@@ -370,7 +350,6 @@ export class Container extends EventEmitterWithErrorHandling<IContainerEvents> i
 
     private readonly logger: ITelemetryLogger;
 
-    private pendingClientId: string | undefined;
     private loaded = false;
     private _attachState = AttachState.Detached;
 
@@ -383,14 +362,12 @@ export class Container extends EventEmitterWithErrorHandling<IContainerEvents> i
         return this._storageService;
     }
 
-    private _clientId: string | undefined;
     private _id: string | undefined;
     private containerUrl: string | undefined;
     private readonly clientDetailsOverride: IClientDetails | undefined;
     private readonly _deltaManager: DeltaManager;
     private _existing: boolean | undefined;
     private service: IDocumentService | undefined;
-    private _connectionState = ConnectionState.Disconnected;
     private readonly _audience: Audience;
 
     private _context: ContainerContext | undefined;
@@ -409,17 +386,15 @@ export class Container extends EventEmitterWithErrorHandling<IContainerEvents> i
     }
 
     private resumedOpProcessingAfterLoad = false;
-    private firstConnection = true;
-    private manualReconnectInProgress = false;
-    private readonly connectionTransitionTimes: number[] = [];
-    private messageCountAfterDisconnection: number = 0;
+
+    private _lastVisible: number | undefined;
     private _loadedFromVersion: IVersion | undefined;
     private _resolvedUrl: IResolvedUrl | undefined;
     private cachedAttachSummary: ISummaryTree | undefined;
     private attachInProgress = false;
     private _dirtyContainer = false;
 
-    private lastVisible: number | undefined;
+    private readonly connectionStateHandler: ConnectionStateHandler;
 
     private _closed = false;
 
@@ -434,6 +409,14 @@ export class Container extends EventEmitterWithErrorHandling<IContainerEvents> i
 
     public get resolvedUrl(): IResolvedUrl | undefined {
         return this._resolvedUrl;
+    }
+
+    public get lastVisible() {
+        return this._lastVisible;
+    }
+
+    public get isLoaded() {
+        return this.loaded;
     }
 
     public get loadedFromVersion(): IVersion | undefined {
@@ -483,11 +466,11 @@ export class Container extends EventEmitterWithErrorHandling<IContainerEvents> i
     }
 
     public get connectionState(): ConnectionState {
-        return this._connectionState;
+        return this.connectionStateHandler.connectionState;
     }
 
     public get connected(): boolean {
-        return this.connectionState === ConnectionState.Connected;
+        return this.connectionStateHandler.connected;
     }
 
     /**
@@ -503,7 +486,7 @@ export class Container extends EventEmitterWithErrorHandling<IContainerEvents> i
      * Set once this.connected is true, otherwise undefined
      */
     public get clientId(): string | undefined {
-        return this._clientId;
+        return this.connectionStateHandler.clientId;
     }
 
     /**
@@ -598,16 +581,23 @@ export class Container extends EventEmitterWithErrorHandling<IContainerEvents> i
         this.logger = ChildLogger.create(this.subLogger, "Container");
 
         this._deltaManager = this.createDeltaManager();
+        this.connectionStateHandler = new ConnectionStateHandler(
+            this,
+            this._deltaManager,
+            this._audience,
+            this.logger,
+            () => this.context,
+            () => this.protocolHandler);
 
         // keep track of last time page was visible for telemetry
         if (typeof document === "object" && document !== null) {
-            this.lastVisible = document.hidden ? performance.now() : undefined;
+            this._lastVisible = document.hidden ? performance.now() : undefined;
             document.addEventListener("visibilitychange", () => {
                 if (document.hidden) {
-                    this.lastVisible = performance.now();
+                    this._lastVisible = performance.now();
                 } else {
                     // settimeout so this will hopefully fire after disconnect event if being hidden caused it
-                    setTimeout(() => this.lastVisible = undefined, 0);
+                    setTimeout(() => this._lastVisible = undefined, 0);
                 }
             });
         }
@@ -641,7 +631,7 @@ export class Container extends EventEmitterWithErrorHandling<IContainerEvents> i
                         }
                         break;
                     case connectEventName:
-                        if (this._connectionState !== ConnectionState.Disconnected) {
+                        if (this.connectionState !== ConnectionState.Disconnected) {
                             listener(event);
                         }
                         break;
@@ -746,9 +736,6 @@ export class Container extends EventEmitterWithErrorHandling<IContainerEvents> i
                 // Get the document state post attach - possibly can just call attach but we need to change the
                 // semantics around what the attach means as far as async code goes.
                 const appSummary: ISummaryTree = this.context.createSummary();
-                if (this.protocolHandler === undefined) {
-                    throw new Error("Protocol Handler is undefined");
-                }
                 const protocolSummary = this.captureProtocolSummary();
                 this.cachedAttachSummary = combineAppAndProtocolSummary(appSummary, protocolSummary);
 
@@ -801,7 +788,7 @@ export class Container extends EventEmitterWithErrorHandling<IContainerEvents> i
             this.cachedAttachSummary = undefined;
 
             // Propagate current connection state through the system.
-            this.propagateConnectionState();
+            this.connectionStateHandler.propagateConnectionState();
             if (!this.closed) {
                 this.resumeInternal({ fetchOpsFromStorage: false, reason: "createDetached" });
             }
@@ -851,9 +838,9 @@ export class Container extends EventEmitterWithErrorHandling<IContainerEvents> i
         // If container state is not attached and resumed, then don't connect to delta stream. Also don't set the
         // manual reconnection flag to true as we haven't made the initial connection yet.
         if (reconnect && this._attachState === AttachState.Attached && this.resumedOpProcessingAfterLoad) {
-            if (this._connectionState === ConnectionState.Disconnected) {
+            if (this.connectionStateHandler.connectionState === ConnectionState.Disconnected) {
                 // Only track this as a manual reconnection if we are truly the ones kicking it off.
-                this.manualReconnectInProgress = true;
+                this.connectionStateHandler.manualReconnectInProgress = true;
             }
 
             // Ensure connection to web socket
@@ -1121,8 +1108,8 @@ export class Container extends EventEmitterWithErrorHandling<IContainerEvents> i
     }
 
     private recordConnectStartTime() {
-        if (this.connectionTransitionTimes[ConnectionState.Disconnected] === undefined) {
-            this.connectionTransitionTimes[ConnectionState.Disconnected] = performance.now();
+        if (this.connectionStateHandler.connectionTransitionTimes[ConnectionState.Disconnected] === undefined) {
+            this.connectionStateHandler.connectionTransitionTimes[ConnectionState.Disconnected] = performance.now();
         }
     }
 
@@ -1213,7 +1200,7 @@ export class Container extends EventEmitterWithErrorHandling<IContainerEvents> i
         await this.loadContext(codeDetails, attributes, snapshot, undefined, pendingLocalState);
 
         // Propagate current connection state through the system.
-        this.propagateConnectionState();
+        this.connectionStateHandler.propagateConnectionState();
 
         if (!pause) {
             this.resume();
@@ -1271,7 +1258,7 @@ export class Container extends EventEmitterWithErrorHandling<IContainerEvents> i
         // The load context - given we seeded the quorum - will be great
         await this.createDetachedContext(attributes);
 
-        this.propagateConnectionState();
+        this.connectionStateHandler.propagateConnectionState();
 
         this.loaded = true;
     }
@@ -1293,7 +1280,7 @@ export class Container extends EventEmitterWithErrorHandling<IContainerEvents> i
 
         this.loaded = true;
 
-        this.propagateConnectionState();
+        this.connectionStateHandler.propagateConnectionState();
     }
 
     private async getDocumentStorageService(): Promise<IDocumentStorageService> {
@@ -1387,13 +1374,7 @@ export class Container extends EventEmitterWithErrorHandling<IContainerEvents> i
             protocolLogger.sendErrorEvent(error);
         });
 
-        // Track membership changes and update connection state accordingly
-        protocol.quorum.on("addMember", (clientId, details) => {
-            // This is the only one that requires the pending client ID
-            if (clientId === this.pendingClientId) {
-                this.setConnectionState(ConnectionState.Connected);
-            }
-        });
+        this.connectionStateHandler.setProtocolMemberEvents(protocol);
 
         protocol.quorum.on("addProposal",(proposal: IPendingProposal) => {
             if (proposal.key === "code" || proposal.key === "code2") {
@@ -1509,46 +1490,6 @@ export class Container extends EventEmitterWithErrorHandling<IContainerEvents> i
             () => this.activeConnection(),
         );
 
-        deltaManager.on(connectEventName, (details: IConnectionDetails, opsBehind?: number) => {
-            const oldState = this._connectionState;
-            this._connectionState = ConnectionState.Connecting;
-
-            // Stash the clientID to detect when transitioning from connecting (socket.io channel open) to connected
-            // (have received the join message for the client ID)
-            // This is especially important in the reconnect case. It's possible there could be outstanding
-            // ops sent by this client, so we should keep the old client id until we see our own client's
-            // join message. after we see the join message for out new connection with our new client id,
-            // we know there can no longer be outstanding ops that we sent with the previous client id.
-            this.pendingClientId = details.clientId;
-
-            this.emit(connectEventName, opsBehind);
-
-            // Report telemetry after we set client id!
-            this.logConnectionStateChangeTelemetry(ConnectionState.Connecting, oldState);
-
-            // Check if we already processed our own join op through delta storage!
-            // we are fetching ops from storage in parallel to connecting to ordering service
-            // Given async processes, it's possible that we have already processed our own join message before
-            // connection was fully established.
-            // Note that we might be still initializing quorum - connection is established proactively on load!
-            if ((this._protocolHandler !== undefined && this._protocolHandler.quorum.has(details.clientId))
-                    || deltaManager.connectionMode === "read") {
-                this.setConnectionState(ConnectionState.Connected);
-            }
-
-            // Back-compat for new client and old server.
-            this._audience.clear();
-
-            for (const priorClient of details.initialClients ?? []) {
-                this._audience.addMember(priorClient.clientId, priorClient.client);
-            }
-        });
-
-        deltaManager.on("disconnect", (reason: string) => {
-            this.manualReconnectInProgress = false;
-            this.setConnectionState(ConnectionState.Disconnected, reason);
-        });
-
         deltaManager.on("throttled", (warning: IThrottlingWarning) => {
             this.raiseContainerWarning(warning);
         });
@@ -1581,131 +1522,6 @@ export class Container extends EventEmitterWithErrorHandling<IContainerEvents> i
             });
     }
 
-    private logConnectionStateChangeTelemetry(
-        value: ConnectionState,
-        oldState: ConnectionState,
-        reason?: string) {
-        // Log actual event
-        const time = performance.now();
-        this.connectionTransitionTimes[value] = time;
-        const duration = time - this.connectionTransitionTimes[oldState];
-
-        let durationFromDisconnected: number | undefined;
-        let connectionMode: string | undefined;
-        let connectionInitiationReason: string | undefined;
-        let autoReconnect: ReconnectMode | undefined;
-        let checkpointSequenceNumber: number | undefined;
-        let sequenceNumber: number | undefined;
-        let opsBehind: number | undefined;
-        if (value === ConnectionState.Disconnected) {
-            autoReconnect = this._deltaManager.reconnectMode;
-        } else {
-            connectionMode = this._deltaManager.connectionMode;
-            sequenceNumber = this.deltaManager.lastSequenceNumber;
-            if (value === ConnectionState.Connected) {
-                durationFromDisconnected = time - this.connectionTransitionTimes[ConnectionState.Disconnected];
-                durationFromDisconnected = TelemetryLogger.formatTick(durationFromDisconnected);
-            } else {
-                // This info is of most interest on establishing connection only.
-                checkpointSequenceNumber = this.deltaManager.lastKnownSeqNumber;
-                if (this.deltaManager.hasCheckpointSequenceNumber) {
-                    opsBehind = checkpointSequenceNumber - sequenceNumber;
-                }
-            }
-            if (this.firstConnection) {
-                connectionInitiationReason = "InitialConnect";
-            } else if (this.manualReconnectInProgress) {
-                connectionInitiationReason = "ManualReconnect";
-            } else {
-                connectionInitiationReason = "AutoReconnect";
-            }
-        }
-
-        this.logger.sendPerformanceEvent({
-            eventName: `ConnectionStateChange_${ConnectionState[value]}`,
-            from: ConnectionState[oldState],
-            duration,
-            durationFromDisconnected,
-            reason,
-            connectionInitiationReason,
-            socketDocumentId: this._deltaManager.socketDocumentId,
-            pendingClientId: this.pendingClientId,
-            clientId: this.clientId,
-            connectionMode,
-            autoReconnect,
-            opsBehind,
-            online: OnlineStatus[isOnline()],
-            lastVisible: this.lastVisible !== undefined ? performance.now() - this.lastVisible : undefined,
-            checkpointSequenceNumber,
-            sequenceNumber,
-        });
-
-        if (value === ConnectionState.Connected) {
-            this.firstConnection = false;
-            this.manualReconnectInProgress = false;
-        }
-    }
-
-    private setConnectionState(value: ConnectionState.Disconnected, reason: string);
-    private setConnectionState(value: ConnectionState.Connecting | ConnectionState.Connected);
-    private setConnectionState(
-        value: ConnectionState,
-        reason?: string) {
-        assert(value !== ConnectionState.Connecting, "Trying to set connection state while container is connecting!");
-        if (this.connectionState === value) {
-            // Already in the desired state - exit early
-            this.logger.sendErrorEvent({ eventName: "setConnectionStateSame", value });
-            return;
-        }
-
-        const oldState = this._connectionState;
-        this._connectionState = value;
-
-        if (value === ConnectionState.Connected) {
-            // Mark our old client should have left in the quorum if it's still there
-            if (this._clientId !== undefined) {
-                const client: ILocalSequencedClient | undefined =
-                    this._protocolHandler?.quorum.getMember(this._clientId);
-                if (client !== undefined) {
-                    client.shouldHaveLeft = true;
-                }
-            }
-            this._clientId = this.pendingClientId;
-        } else if (value === ConnectionState.Disconnected) {
-            // Important as we process our own joinSession message through delta request
-            this.pendingClientId = undefined;
-        }
-
-        if (this.loaded) {
-            this.propagateConnectionState();
-        }
-
-        // Report telemetry after we set client id!
-        this.logConnectionStateChangeTelemetry(value, oldState, reason);
-    }
-
-    private propagateConnectionState() {
-        const logOpsOnReconnect: boolean =
-            this._connectionState === ConnectionState.Connected &&
-            !this.firstConnection &&
-            this._deltaManager.connectionMode === "write";
-        if (logOpsOnReconnect) {
-            this.messageCountAfterDisconnection = 0;
-        }
-
-        const state = this._connectionState === ConnectionState.Connected;
-        if (!this.context.disposed) {
-            this.context.setConnectionState(state, this.clientId);
-        }
-        this.protocolHandler.quorum.setConnectionState(state, this.clientId);
-        raiseConnectedEvent(this.logger, this, state, this.clientId);
-
-        if (logOpsOnReconnect) {
-            this.logger.sendTelemetryEvent(
-                { eventName: "OpsSentOnReconnect", count: this.messageCountAfterDisconnection });
-        }
-    }
-
     private submitContainerMessage(type: MessageType, contents: any, batch?: boolean, metadata?: any): number {
         const outboundMessageType: string = type;
         switch (outboundMessageType) {
@@ -1726,7 +1542,7 @@ export class Container extends EventEmitterWithErrorHandling<IContainerEvents> i
             return -1;
         }
 
-        this.messageCountAfterDisconnection += 1;
+        this.connectionStateHandler.messageCountAfterDisconnection += 1;
         this.collabWindowTracker.stopSequenceNumberUpdate();
         return this._deltaManager.submit(type, contents, batch, metadata);
     }
@@ -1747,7 +1563,7 @@ export class Container extends EventEmitterWithErrorHandling<IContainerEvents> i
                 const error = new DataCorruptionError(
                     errorMsg,
                     {
-                        clientId: this._clientId,
+                        clientId: this.clientId,
                         messageClientId: message.clientId,
                         sequenceNumber: message.sequenceNumber,
                         clientSequenceNumber: message.clientSequenceNumber,
@@ -1758,7 +1574,7 @@ export class Container extends EventEmitterWithErrorHandling<IContainerEvents> i
             }
         }
 
-        const local = this._clientId === message.clientId;
+        const local = this.clientId === message.clientId;
 
         // Forward non system messages to the loaded runtime for processing
         if (!isSystemMessage(message)) {
@@ -1790,7 +1606,7 @@ export class Container extends EventEmitterWithErrorHandling<IContainerEvents> i
                 this._audience.removeMember(leftClientId);
             }
         } else {
-            const local = this._clientId === message.clientId;
+            const local = this.clientId === message.clientId;
             this.context.processSignal(message, local);
         }
     }
