+++ packages/loader/container-loader/src/index.ts
@@ -10,3 +10,4 @@ export * from "./loader";
 export * from "./networkUtils";
 export * from "./utils";
 export * from "./retriableDocumentStorageService";
+export * from "./connectionStateHandler";
